Refactor student and subject pages with Dynamic path routes  & Login integration
Moved inline styles from student and subject PHP files to a new style.css for better maintainability. Updated file paths and form actions to use relative links and centralized routing via index.php. Improved deletion logic for grades and students, implemented soft delete for students, and added duplicate checks with alert messages for student and subject creation. Cleaned up code and removed unnecessary session usage.

+++ grade/delete.php
@@ -3,15 +3,22 @@
 	
 	require_once ('../config.php');
 		
-	$query="DELETE FROM grades WHERE gr_id='$id' ";
-		
-	$results = mysqli_query($connect,$query);
-					
-	if(!$results){
-		echo mysqli_error($connect);
-	}else{
-		echo "query Executed !!";
+	$check_query = "SELECT grade_id FROM students WHERE grade_id='$id'";
+	$check_result=mysqli_query($connect,$check_query);
+	$row= mysqli_num_rows($check_result);
+
+	if($row > 0){
+		echo "Can't Delete! Grade already in students record ";
+	} else {
+		$query="DELETE FROM grades WHERE gr_id='$id' ";
+			
+		$results = mysqli_query($connect,$query);
+						
+		if(!$results){
+			echo mysqli_error($connect);
+		}else{
+			header("location:../index.php?section=grade&page=index");
+		}
 	}
 	
-	header("location:../index.php?section=grade&page=index");
 ?>
